+++ scripts/create_productive_database_from_dumps.sh
@@ -6,7 +6,36 @@ source "$folder_name/configuration.sh"
 source "$folder_name/helper_functions.sh"
 
 # create new productive database from dumps
-# delete old one if available
+echo -e "\nCreate new productive database from dump -- started at $(get_timestamp)"
+
+# check transfered files
+if [ ! -f "$dumped_tables_file" ]; then
+    echo "dumped_tables file missing"
+    exit 23
+fi
+temp_db_map_state_file="$temp_folder/$(basename $productive_db_map_state_file)"
+if [ ! -f "$temp_db_map_state_file" ]; then
+    echo "temp state file missing"
+    exit 23
+fi
+md5_file="$dumped_tables_file.md5"
+if [ ! -f "$md5_file" ]; then
+    echo "md5 file missing"
+    exit 23
+fi
+
+# check checksum
+old_directory=$(pwd)
+cd "$temp_folder"
+md5sum -c "$md5_file"
+if [[ $? != 0 ]]; then
+    echo "Checksums don't match"
+    cd "$old_directory"
+    exit 21
+fi
+cd "$old_directory"
+
+# delete old temp database if available
 result=$(psql -h $server_address -U $user_name -l | grep -i "$db_tmp_name ")
 if [ ! -z "$result" ]; then
     psql -h $server_address -U $user_name -d postgres -c "DROP DATABASE $db_tmp_name;"
@@ -15,26 +44,16 @@ if [ ! -z "$result" ]; then
     fi
 fi
 
-# copy map state file to provide the productive db version information
-cp "$map_state_file" "$productive_db_map_state_file"
-
 # create new db from raw db template
-echo -e "\nCreate new temp osm database -- started at $(get_timestamp)"
-createdb -h $server_address -U $user_name -O $user_name -T "$db_raw_name" "$db_tmp_name"
+createdb -h $server_address -U $user_name -O $user_name "$db_tmp_name"
 if [[ $? != 0 ]]; then
     echo "Can't create temp database from raw database template"
     exit 22
 fi
 
-# check for dump_tables file
-if [ ! -f "$dumped_tables_file" ]; then
-    echo "dumped_tables file missing"
-    exit 23
-fi
-
 # import into temp database
 echo -e "\nImport dumped tables into database -- started at $(get_timestamp)"
-pg_restore -h $server_address -U $user_name -d $db_tmp_name -e -1 $dumped_tables_file
+pg_restore -h $server_address -p $server_port -U $user_name -d "$db_tmp_name" -e -j2 "$dumped_tables_file"
 if [[ $? != 0 ]]; then
     echo "Can't import dumped tables into temp database"
     exit 24
@@ -85,6 +104,13 @@ psql -h $server_address -U $user_name -d postgres -c "ALTER DATABASE $db_tmp_nam
 if [[ $? != 0 ]]; then
     echo -e "\nCan't rename temp database to active one"
     exit 27
+fi
+
+# copy map state file to provide the productive db version information
+cp "$temp_db_map_state_file" "$productive_db_map_state_file"
+if [[ $? != 0 ]]; then
+    echo "Can't copy map state file"
+    exit 28
 fi
 
 echo -e "\nProductive database created from dumps at $(get_timestamp)"
